fix: limit the course runs in course page to the ones in catalog
By default, the course detail page lists all the available course runs
of a course for the learner to enroll. However, this exposes course runs
that are not included in a enterprise catalog to the learners. In this
commit, the course runs are filtered to allow only the runs mentioned
in the catalog to be shown to the learner.

Internal-ref: https://tasks.opencraft.com/browse/BB-9491